Add 'interval' parameter to crypto_history
Hello! just a casual stumbler by. Went to use this function and noticed it didn't have the 'interval' parameter for more coarsely sampling historical prices. Not sure about your style guidelines, but tried to match them! 

- added to docstring
- added arg with default = NULL
- added check for input, defaulting to 'daily', and then checking against the list of allowable values to declare it 'daily' if an invalid value was passed.

should behave identically to if 'interval' wasn't passed, but give the option if people want it. Feel free to reject if you don't want it, no worries :)